Working app without tests
