+++ src/main/java/com/vyacheslavbasovproject/developerapitask/controller/exceptionsHandling/ValidationErrorResponse.java
@@ -1,0 +1,15 @@
+package com.vyacheslavbasovproject.developerapitask.controller.exceptionsHandling;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ValidationErrorResponse {
+    private List<Violation> violations = new ArrayList<>();
+}
