+++ src/main/java/com/vyacheslavbasovproject/developerapitask/controller/exceptionsHandling/Violation.java
@@ -1,0 +1,14 @@
+package com.vyacheslavbasovproject.developerapitask.controller.exceptionsHandling;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Violation {
+    private String fieldName;
+
+    private String message;
+}
